+++ apps/ledger-live-mobile/src/newArch/features/LandingPages/screens/LargeMoverLandingPage/components/LargeMoverGraph.tsx
@@ -1,0 +1,52 @@
+import React, { useCallback, useMemo } from "react";
+import { ensureContrast } from "@ledgerhq/native-ui";
+import { Item } from "~/components/Graph/types";
+import Graph from "~/components/Graph";
+import { KeysPriceChange, MarketCoinDataChart } from "@ledgerhq/live-common/market/utils/types";
+import { CryptoOrTokenCurrency } from "@ledgerhq/types-cryptoassets";
+import { useTheme } from "styled-components/native";
+import { getCryptoCurrencyById, getCurrencyColor } from "@ledgerhq/live-common/currencies/index";
+
+type GraphProps = {
+  chartData: MarketCoinDataChart;
+  range: KeysPriceChange;
+  currencyId: string;
+  width: number;
+};
+
+export const LargeMoverGraph: React.FC<GraphProps> = ({ chartData, range, currencyId, width }) => {
+  const theme = useTheme();
+  const currency: CryptoOrTokenCurrency | undefined = getCryptoCurrencyById(currencyId);
+
+  const data = useMemo(
+    () =>
+      range && chartData?.[range]
+        ? chartData[range].map(d => ({
+            date: new Date(d[0]),
+            value: d[1] || 0,
+          }))
+        : [],
+    [chartData, range],
+  );
+
+  const mapGraphValue = useCallback((d: Item) => d?.value ?? 0, []);
+
+  const graphColor = useMemo(
+    () =>
+      !currency
+        ? theme.colors.primary.c80
+        : ensureContrast(getCurrencyColor(currency), theme.colors.background.main),
+    [theme.colors.background.main, theme.colors.primary.c80, currency],
+  );
+  return (
+    <Graph
+      testID="large-mover-graph"
+      width={width}
+      height={150}
+      color={graphColor}
+      isInteractive={false}
+      mapValue={mapGraphValue}
+      data={data}
+    />
+  );
+};