+++ apps/ledger-live-mobile/src/screens/Account/ListHeaderComponent.tsx
@@ -88,6 +88,7 @@ export function useListHeaderComponents({
   listHeaderComponents: ReactNode[];
   stickyHeaderIndices?: number[];
 } {
+  const llmNftSupport = useFeature("llNftSupport");
   const llmSolanaNfts = useFeature("llmSolanaNfts");
   if (!account) return { listHeaderComponents: [], stickyHeaderIndices: undefined };
 
@@ -134,7 +135,8 @@ export function useListHeaderComponents({
     isStuckOperation({ family: mainAccount.currency.family, operation: oldestEditableOperation });
 
   const displayNftCollections = isNFTCollectionsDisplayable(account, empty, {
-    llmSolanaNftsEnabled: llmSolanaNfts?.enabled,
+    llmNftSupportEnabled: !!llmNftSupport?.enabled,
+    llmSolanaNftsEnabled: !!llmSolanaNfts?.enabled,
   });
 
   const disableDelegation =