+++ libs/coin-modules/coin-polkadot/jest.config.js
@@ -11,7 +11,11 @@ module.exports = {
   coverageReporters: ["json", ["lcov", { file: "polkadot-lcov.info", projectRoot: "../" }], "text"],
   testEnvironment: "node",
   testPathIgnorePatterns: ["lib/", "lib-es/", ".*\\.(integ|integration)\\.test\\.[tj]s"],
-  setupFilesAfterEnv: ["jest-expect-message", "dotenv/config"],
+  setupFilesAfterEnv: [
+    "jest-expect-message",
+    "dotenv/config",
+    // "@ledgerhq/disable-network-setup"
+  ],
   reporters: [
     "default",
     [