+++ libs/coin-modules/coin-solana/jest.config.js
@@ -32,4 +32,5 @@ module.exports = {
       { outputName: "solana-sonar-executionTests-report.xml", reportedFilePath: "absolute" },
     ],
   ],
+  setupFilesAfterEnv: ["@ledgerhq/disable-network-setup"],
 };