+++ libs/ui/packages/react/src/pre-ldls/components/AssetList/AssetList.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useCallback } from "react";
 import { AssetItem, AssetType } from "../AssetItem/AssetItem";
 import { VirtualList } from "../VirtualList/VirtualList";
 
@@ -9,13 +9,10 @@ export const AssetList = ({
   assets: AssetType[];
   onClick: (asset: AssetType) => void;
 }) => {
-  return (
-    <VirtualList
-      itemHeight={64}
-      count={assets.length}
-      renderItem={(i: number) => (
-        <AssetItem name={assets[i].name} ticker={assets[i].ticker} onClick={onClick} />
-      )}
-    />
+  const renderAssetItem = useCallback(
+    ({ name, ticker }: AssetType) => <AssetItem name={name} ticker={ticker} onClick={onClick} />,
+    [onClick],
   );
+
+  return <VirtualList itemHeight={64} items={assets} renderItem={renderAssetItem} />;
 };