+++ libs/ledgerjs/packages/hw-app-celo/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ledgerhq/hw-app-polkadot
 
+## 6.33.2-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-app-eth@6.45.4-next.0
+
 ## 6.33.1
 
 ### Patch Changes