+++ libs/ui/examples/next.js/CHANGELOG.md
@@ -1,5 +1,12 @@
 # next.js-example
 
+## 0.1.24-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`41d82c2`](https://github.com/LedgerHQ/ledger-live/commit/41d82c211ffc03882d101fecd1175d8f8a823341), [`98adc70`](https://github.com/LedgerHQ/ledger-live/commit/98adc70de21ac0423e378ea2a7076e15e8fc8191), [`e0cf0ed`](https://github.com/LedgerHQ/ledger-live/commit/e0cf0ed7987da2e4cd7cfcc389e9804ab86129f9)]:
+  - @ledgerhq/react-ui@0.25.0-next.0
+
 ## 0.1.23
 
 ### Patch Changes